final/backend/requirements.txt
fastapi
uvicorn
python-socketio[asyncio]
httpx>=0.27
openai
python-dotenv
pandas
openpyxl
beautifulsoup4>=4.12
lxml>=5.0
tldextract>=5.1
# optional (recommended for richer link extraction)
trafilatura>=1.7
pdfminer.six
pdfplumber
redis>=5.0
# LangChain & LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20
rank-bm25>=0.2.2  # BM25 retriever 의존성